Move solvers into their own files.
